requirements.txt
grpcio
grpcio-tools
pika==1.3.2
fastapi==0.109.0
uvicorn==0.27.0
streamlit
requests